+++ oap-server/analyzer/meter-analyzer/src/main/java/org/apache/skywalking/oap/meter/analyzer/dsl/DSL.java
@@ -28,6 +28,7 @@ import java.util.Map;
 
 import org.apache.skywalking.oap.meter.analyzer.dsl.registry.ProcessRegistry;
 import org.apache.skywalking.oap.meter.analyzer.dsl.tagOpt.K8sRetagType;
+import org.apache.skywalking.oap.meter.analyzer.k8s.Kubernetes;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.core.source.DetectPoint;
 import org.codehaus.groovy.ast.stmt.DoWhileStatement;
@@ -57,6 +58,7 @@ public final class DSL {
         icz.addImport("DetectPoint", DetectPoint.class.getName());
         icz.addImport("Layer", Layer.class.getName());
         icz.addImport("ProcessRegistry", ProcessRegistry.class.getName());
+        icz.addImport("Kubernetes", Kubernetes.class.getName());
         cc.addCompilationCustomizers(icz);
 
         final SecureASTCustomizer secureASTCustomizer = new SecureASTCustomizer();
@@ -77,6 +79,7 @@ public final class DSL {
                          .add(DetectPoint.class)
                          .add(Layer.class)
                          .add(ProcessRegistry.class)
+                         .add(Kubernetes.class)
                 .build());
         cc.addCompilationCustomizers(secureASTCustomizer);