+++ oap-server/server-receiver-plugin/skywalking-zabbix-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zabbix/provider/config/ZabbixConfig.java
@@ -30,6 +30,7 @@ public class ZabbixConfig implements MetricRuleConfig {
     private String expSuffix;
     private String expPrefix;
     private String filter;
+    private String initExp;
     private Entities entities;
     private List<String> requiredZabbixItemKeys;
     private List<Metric> metrics;